refactor: ♻️ detect hmr
